chore: rename it-tests/ to ui-tests/ and restructure internals
- Rename it-tests/ → ui-tests/ (aligns with Jenkinsfile 'UI Tests' stage)
- Create ui-tests/editor/ for the six root-level loose test files
- Move BasicFlow, ContextualMenuOpen, MavenDependencyUpdate,
  OpenTextualEditorCommand, PropertyPanelLoading, SwitchBetweenTabs
  into ui-tests/editor/ and fix their Util/pageObjects import paths
- Rename tsconfig.it-tests.json → tsconfig.ui-tests.json
- Update all config files: eslint.config.mjs, extester.config.json,
  .vscodeignore, .vscode/settings.json
- Rename scripts: build:test:it → build:test:ui, test:it* → test:ui*,
  update all ./it-tests/ path refs in package.json
- Update Jenkinsfile: test:it:with-prebuilt-vsix → test:ui:with-prebuilt-vsix
- Update .github/workflows/_integration-tests.yaml (all 3 platforms)
- Update AGENTS.md and CONTRIBUTING.md
