+++ .github/workflows/_integration-tests.yaml
@@ -136,7 +136,7 @@ jobs:
           timeout_minutes: 30
           retry_wait_seconds: ${{ inputs.retry-wait-macos }}
           max_attempts: 2
-          command: yarn run test:it:with-prebuilt-vsix
+          command: yarn run test:ui:with-prebuilt-vsix
           on_retry_command: rm -rf "${RUNNER_TEMP}/extest-code"
       - uses: ./.github/actions/upload-test-artifacts
         if: failure()
@@ -174,7 +174,7 @@ jobs:
           max_attempts: 2
           command: |
             Set-DisplayResolution -Width 1920 -Height 1080 -Force
-            yarn run test:it:with-prebuilt-vsix
+            yarn run test:ui:with-prebuilt-vsix
           on_retry_command: Remove-Item -Recurse -Force "$env:RUNNER_TEMP\extest-code"
       - uses: ./.github/actions/upload-test-artifacts
         if: failure()
