+++ package.json
@@ -18,10 +18,10 @@
     "email": "apupier@redhat.com"
   },
   "scripts": {
-    "lint": "eslint src it-tests",
+    "lint": "eslint src ui-tests",
     "build:prod": "rimraf dist && yarn run compile && yarn run lint",
     "build:dev": "rimraf dist && yarn run watch && yarn run lint",
-    "build:test:it": "rimraf ./out && tsc --project tsconfig.it-tests.json --skipLibCheck --sourceMap true && yarn run lint",
+    "build:test:ui": "rimraf ./out && tsc --project tsconfig.ui-tests.json --skipLibCheck --sourceMap true && yarn run lint",
     "build:test:unit": "rimraf ./out && tsc --project tsconfig.unit-tests.json --skipLibCheck --sourceMap true && yarn run lint",
     "build:vsix": "yarn vsce package --no-dependencies --yarn",
     "compile": "webpack",
@@ -29,11 +29,11 @@
     "watch": "webpack --env dev",
     "run:webmode": "yarn vscode-test-web --browserType=chromium --extensionDevelopmentPath=. --open-devtools ./resources",
     "test:unit": "yarn build:test:unit && vscode-test",
-    "test:it": "yarn build:test:it && extest setup-and-run",
-    "setup:test:it:with-prebuilt-vsix": "yarn build:test:it && extest get-vscode && extest get-chromedriver && extest install-vsix --vsix_file vscode-kaoto-$npm_package_version.vsix --extensions_dir ./test-resources && extest run-tests --uninstall_extension --extensions_dir ./test-resources --open_resource './test Fixture with speci@l chars' --custom_page_objects './out/pageObjects/locators.js'",
-    "test:it:with-prebuilt-vsix": "yarn setup:test:it:with-prebuilt-vsix --code_settings ./it-tests/vscode-settings.json 'out/**/*.test.js'",
-    "test:it:with-prebuilt-vsix:minikube": "yarn setup:test:it:with-prebuilt-vsix --code_settings ./it-tests/vscode-settings-minikube.json 'out/**/*.test.js'",
-    "test:it:clean": "rimraf ./test-resources && rimraf ./out && rimraf *.vsix"
+    "test:ui": "yarn build:test:ui && extest setup-and-run",
+    "setup:test:ui:with-prebuilt-vsix": "yarn build:test:ui && extest get-vscode && extest get-chromedriver && extest install-vsix --vsix_file vscode-kaoto-$npm_package_version.vsix --extensions_dir ./test-resources && extest run-tests --uninstall_extension --extensions_dir ./test-resources --open_resource './test Fixture with speci@l chars' --custom_page_objects './out/pageObjects/locators.js'",
+    "test:ui:with-prebuilt-vsix": "yarn setup:test:ui:with-prebuilt-vsix --code_settings ./ui-tests/vscode-settings.json 'out/**/*.test.js'",
+    "test:ui:with-prebuilt-vsix:minikube": "yarn setup:test:ui:with-prebuilt-vsix --code_settings ./ui-tests/vscode-settings-minikube.json 'out/**/*.test.js'",
+    "test:ui:clean": "rimraf ./test-resources && rimraf ./out && rimraf *.vsix"
   },
   "dependencies": {
     "@kaoto/camel-catalog": "^0.8.2",
