+++ tsconfig.ui-tests.json
@@ -3,7 +3,7 @@
   "compilerOptions": {
     "module": "CommonJS",
     "moduleResolution": "Bundler",
-    "rootDir": "it-tests",
+    "rootDir": "ui-tests",
     "target": "ES2023",
     "lib": ["ES2023"],
     "strict": true,
@@ -14,5 +14,5 @@
     "resolveJsonModule": true,
     "outDir": "out"
   },
-  "include": ["it-tests"]
+  "include": ["ui-tests"]
 }
