+++ ui-tests/editor/OpenTextualEditorCommand.test.ts
