+++ ui-tests/editor/PropertyPanelLoading.test.ts
@@ -15,8 +15,8 @@
  */
 import { EditorView, VSBrowser, WebDriver, WebView } from 'vscode-extension-tester';
 import * as path from 'path';
-import { clickWhenClickable, dismissHoverOverlay, openAndSwitchToKaotoFrame } from './Util';
-import { KaotoCanvas, KaotoEditor } from './pageObjects';
+import { clickWhenClickable, dismissHoverOverlay, openAndSwitchToKaotoFrame } from '../Util';
+import { KaotoCanvas, KaotoEditor } from '../pageObjects';
 
 describe('Property panel loading test', function () {
 	this.timeout(60_000);
