+++ ui-tests/pageObjects/CatalogModal.ts
