+++ ui-tests/pageObjects/EditorTabs.ts
