+++ ui-tests/pageObjects/KaotoCanvas.ts
