+++ ui-tests/views/07_DeploymentsView.test.ts
