+++ ui-tests/views/08_DeploymentsViewLogsAndStop.test.ts
