+++ ui-tests/views/09_DeploymentsViewRoutesManipulaton.test.ts
