+++ ui-tests/vscode-settings-minikube.json
